+++ src/fnschoo1/canteen/urls.py
@@ -62,11 +62,6 @@ urlpatterns = [
         name="new_consumption",
     ),
     path(
-        "new_consumption/<int:consumption_id>",
-        views.new_consumption,
-        name="new_consumption",
-    ),
-    path(
         "create_consumptions",
         views.create_consumptions,
         name="create_consumptions",