+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 
 # Changelog
 
+## [20260607.80137.820] - 2026-06-07
+### Added
+
+### Changed
+
+### Deprecated
+
+### Fixed
+- `Canteen/Consumption` page: Set the minimum consumption quantity to 0, and ensure it cannot be negative.
+- `Canteen/Consumptions` page: Fixed the issue where updating the consumption quantity failed after deleting the existing consumption quantity.
+- `Canteen/Consumptions` page: Fixed the issue of mismatch between the header and columns of the Consumptions Table.
+### Removed
+
 ## [20260601.81344.845] - 2026-06-01
 ### Added
 - Search for ingredients: The input box allows for searching by quantity and total price.